requirements.txt
tensorflow~=2.14.0
numpy~=1.26.1
sqlalchemy~=2.0.22

matplotlib~=3.8.0
requests~=2.31.0
django-allauth==0.49.0
django-ckeditor==6.2.0
django-autoslug==1.9.8
django-modeltranslation==0.17.5
django-rosetta==0.9.4
Collectfast==2.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-hosts==5.2

Django~=4.2.6
django-mptt==0.13.4
django-currentuser==0.5.3